+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes from [upstream](https://github.com/programble/omgircd).
 * Rename this fork omgircd3
 * Build a package and upload it on PyPI.
 * Logging level is now configurable (via an environment variable)
+* Added debug messages for User events.
 
 ## Roadmap